# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AB Sample API",
    "version": "1.0.0",
    "description": "UX/A-B experiment platform API with analytics and commerce chat agents."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/collect": {
      "post": {
        "summary": "Collect client events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "events": {
                    "type": "array",
                    "items": { "type": "object", "additionalProperties": true }
                  }
                },
                "required": ["events"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Events accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "received": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/experiments": {
      "get": {
        "summary": "List experiments",
        "parameters": [
          { "name": "site_id", "in": "query", "schema": { "type": "string", "default": "ab-sample" } }
        ],
        "responses": {
          "200": { "description": "Experiment list" }
        }
      }
    },
...
    "/api/config": {
      "get": {
        "summary": "Runtime experiment config for SDK",
        "parameters": [
          { "name": "site_id", "in": "query", "schema": { "type": "string", "default": "ab-sample" } },
          { "name": "url", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Config result" }
        }
      }
    }
  }
}